Windows, bzip2.h: Include minwindef.h instead of windows.h
The public header includes windows.h, which is not always desirable and
does not appear to be required. Only the WINAPI definition is required,
for the BZ API calling convention.

Switch from windows.h to wtypes.h + minwindef.h, which define WINAPI.
wtypes.h is required because only including minwindef.h brings in
winnt.h which causes problems.

Resolves: https://gitlab.com/bzip2/bzip2/-/issues/37